move flag struct to remove dependency on kingpinflag
